+++ js/util-const.js
@@ -1,0 +1,38 @@
+
+const NAMES = [
+  'Андрей',
+  'Алексей',
+  'Дмитрий',
+  'Илья',
+  'Евгений',
+  'Екатерина',
+  'Ксения',
+  'Юрий',
+  'Карина',
+  'Татьяна'
+];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+'Супер!',
+'Вау, это круто',
+'Мне кажется это красиво',
+'Я художник, я так вижу',
+'Ну такоое',
+'Вот',
+'Что было то и выложил',
+'норм фотка',
+'и так сойдёт'
+]
+const minLikesCount = 15;
+const maxLikesCount = 200;
+const commentsCount = 5;
+const constValue = 25;
+export {NAMES, MESSAGES, DESCRIPTIONS, minLikesCount, maxLikesCount, commentsCount, constValue}
